refactor(web): consolidate request logging admin UI
